LayoutTests/fast/selectors/nth-child-of-basics-expected.txt

Check the basic features of the ":nth-child(An+B of selectorList)" pseudo class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":nth-child(2n of *)"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)').length is 15
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[4].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[5].id is "testcase12"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[6].id is "testcase14"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[7].id is "testcase16"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[8].id is "testcase18"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[9].id is "testcase20"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[10].id is "testcase22"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[11].id is "testcase24"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[12].id is "testcase26"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[13].id is "testcase28"
PASS document.querySelectorAll('#test-root :nth-child(2n of *)')[14].id is "testcase30"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10","testcase12","testcase14","testcase16","testcase18","testcase20","testcase22","testcase24","testcase26","testcase28","testcase30"]

Testing ":nth-child(2n + 1 of testcaseA)"
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)').length is 8
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)')[1].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)')[2].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)')[3].id is "testcase13"
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)')[4].id is "testcase17"
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)')[5].id is "testcase21"
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)')[6].id is "testcase25"
PASS document.querySelectorAll('#test-root :nth-child(2n + 1 of testcaseA)')[7].id is "testcase29"
PASS styledElements is ["testcase1","testcase5","testcase9","testcase13","testcase17","testcase21","testcase25","testcase29"]

Testing ":nth-child(3n - 2 of [attribute1=value1])"
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])').length is 8
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])')[1].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])')[2].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])')[3].id is "testcase13"
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])')[4].id is "testcase17"
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])')[5].id is "testcase21"
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])')[6].id is "testcase24"
PASS document.querySelectorAll('#test-root :nth-child(3n - 2 of [attribute1=value1])')[7].id is "testcase28"
PASS styledElements is ["testcase1","testcase6","testcase10","testcase13","testcase17","testcase21","testcase24","testcase28"]

Testing ":nth-child(4n - 1 of .bar)"
PASS document.querySelectorAll('#test-root :nth-child(4n - 1 of .bar)').length is 5
PASS document.querySelectorAll('#test-root :nth-child(4n - 1 of .bar)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(4n - 1 of .bar)')[1].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(4n - 1 of .bar)')[2].id is "testcase15"
PASS document.querySelectorAll('#test-root :nth-child(4n - 1 of .bar)')[3].id is "testcase20"
PASS document.querySelectorAll('#test-root :nth-child(4n - 1 of .bar)')[4].id is "testcase25"
PASS styledElements is ["testcase4","testcase10","testcase15","testcase20","testcase25"]

Testing ":nth-child(2n + 2 of :empty)"
PASS document.querySelectorAll('#test-root :nth-child(2n + 2 of :empty)').length is 2
PASS document.querySelectorAll('#test-root :nth-child(2n + 2 of :empty)')[0].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(2n + 2 of :empty)')[1].id is "testcase27"
PASS styledElements is ["testcase9","testcase27"]

Testing "testcaseA:nth-child(even of testcaseA)"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA)').length is 7
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA)')[1].id is "testcase7"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA)')[2].id is "testcase11"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA)')[3].id is "testcase15"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA)')[4].id is "testcase19"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA)')[5].id is "testcase23"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA)')[6].id is "testcase27"
PASS styledElements is ["testcase3","testcase7","testcase11","testcase15","testcase19","testcase23","testcase27"]

Testing "testcaseA:nth-child(even of testcaseA.bazoo)"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA.bazoo)').length is 6
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA.bazoo)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA.bazoo)')[1].id is "testcase7"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA.bazoo)')[2].id is "testcase13"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA.bazoo)')[3].id is "testcase17"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA.bazoo)')[4].id is "testcase23"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of testcaseA.bazoo)')[5].id is "testcase29"
PASS styledElements is ["testcase3","testcase7","testcase13","testcase17","testcase23","testcase29"]

Testing "[attribute2]:nth-child(odd of [attribute2])"
PASS document.querySelectorAll('#test-root [attribute2]:nth-child(odd of [attribute2])').length is 3
PASS document.querySelectorAll('#test-root [attribute2]:nth-child(odd of [attribute2])')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root [attribute2]:nth-child(odd of [attribute2])')[1].id is "testcase9"
PASS document.querySelectorAll('#test-root [attribute2]:nth-child(odd of [attribute2])')[2].id is "testcase20"
PASS styledElements is ["testcase2","testcase9","testcase20"]

Testing ".foo:nth-child(even of .foo)"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)').length is 11
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[1].id is "testcase6"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[2].id is "testcase9"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[3].id is "testcase11"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[4].id is "testcase14"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[5].id is "testcase16"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[6].id is "testcase19"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[7].id is "testcase23"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[8].id is "testcase25"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[9].id is "testcase28"
PASS document.querySelectorAll('#test-root .foo:nth-child(even of .foo)')[10].id is "testcase30"
PASS styledElements is ["testcase4","testcase6","testcase9","testcase11","testcase14","testcase16","testcase19","testcase23","testcase25","testcase28","testcase30"]

Testing "testcaseA:nth-child(even of .bar.foo)"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of .bar.foo)').length is 2
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of .bar.foo)')[0].id is "testcase21"
PASS document.querySelectorAll('#test-root testcaseA:nth-child(even of .bar.foo)')[1].id is "testcase25"
PASS styledElements is ["testcase21","testcase25"]

Testing "*:nth-child(3n+1 of [id][attribute2^="value"].bazoo)"
PASS document.querySelectorAll('#test-root *:nth-child(3n+1 of [id][attribute2^="value"].bazoo)').length is 2
PASS document.querySelectorAll('#test-root *:nth-child(3n+1 of [id][attribute2^="value"].bazoo)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root *:nth-child(3n+1 of [id][attribute2^="value"].bazoo)')[1].id is "testcase20"
PASS styledElements is ["testcase2","testcase20"]

Testing ":-webkit-any(testcaseA, dontExist):nth-child(n+3 of .foo:empty)"
PASS document.querySelectorAll('#test-root :-webkit-any(testcaseA, dontExist):nth-child(n+3 of .foo:empty)').length is 1
PASS document.querySelectorAll('#test-root :-webkit-any(testcaseA, dontExist):nth-child(n+3 of .foo:empty)')[0].id is "testcase27"
PASS styledElements is ["testcase27"]

Testing "*:nth-child(even of :nth-child(odd of [attribute2]))"
PASS document.querySelectorAll('#test-root *:nth-child(even of :nth-child(odd of [attribute2]))').length is 1
PASS document.querySelectorAll('#test-root *:nth-child(even of :nth-child(odd of [attribute2]))')[0].id is "testcase9"
PASS styledElements is ["testcase9"]

PASS successfullyParsed is true

TEST COMPLETE
